+++ TombEngine/Scripting/Internal/TEN/Flow/Enums/FreezeModes.h
@@ -9,19 +9,14 @@ namespace TEN::Scripting
 	// @enum Flow.FreezeMode
 	// @pragma nostrip
 
-	/// Flow.FreezeMode constants.
-	// 
-	// The following constants are inside Flow.FreezeMode.
-	// 
-	//	FreezeMode.NONE - Normal in-game operation.
-	//	FreezeMode.FULL - Game is completely frozen, as in pause or inventory menus.
-	//	FreezeMode.SPECTATOR - Game is completely frozen, but with ability to control camera.
-	//	FreezeMode.PLAYER - Game is completely frozen, but with ability to control player. Experimental.
-	// 
-	// @section Flow.FreezeMode
-
-	/// Table of freeze modes.
+	/// Table of Flow.FreezeMode constants.
+	// To be used with @{Flow.GetFreezeMode} and @{Flow.SetFreezeMode} functions.
 	// @table FreezeMode
+	// 
+	// - `NONE` - Normal in-game operation.
+	// - `FULL` - Game is completely frozen, as in pause or inventory menus.
+	// - `SPECTATOR` - Game is completely frozen, but with ability to control camera.
+	// - `PLAYER` - Game is completely frozen, but with ability to control player. Experimental.
 
 	static const auto FREEZE_MODES = std::unordered_map<std::string, FreezeMode>
 	{