+++ TombEngine/Scripting/Internal/TEN/Flow/Settings/Settings.cpp
@@ -2,7 +2,7 @@
 #include "Scripting/Internal/TEN/Flow/Settings/Settings.h"
 
 #include "Game/effects/Hair.h"
-#include "Scripting/Internal/TEN/Flow/Enums/WeaponTypes.h"
+#include "Scripting/Internal/TEN/Objects/Lara/WeaponTypes.h"
 #include "Scripting/Internal/ReservedScriptNames.h"
 #include "Scripting/Internal/ScriptUtil.h"