+++ TombEngine/Scripting/Internal/TEN/Strings/DisplayString/DisplayString.h
@@ -10,16 +10,13 @@ Constants for Display String Options.
 @pragma nostrip
 */
 
-/*** Strings.DisplayStringOption constants.
-
+/*** Strings.DisplayStringOption constants. To be used with @{Strings.DisplayString} class.
 @table Strings.DisplayStringOption
-	CENTER	-- set the horizontal origin point to the center of the string.
-	RIGHT	-- set the horizontal origin point to right of the string.
-	SHADOW	-- gives the string a small drop shadow.
-	BLINK	-- blinks the string
-*/
-/***
 
+ - `CENTER` - set the horizontal origin point to the center of the string.
+ - `RIGHT` - set the horizontal origin point to right of the string.
+ - `SHADOW` - gives the string a small drop shadow.
+ - `BLINK` - blinks the string
 */
 
 enum class DisplayStringOptions